adding a type-tag to all classes
git-svn-id: http://pinocchio.unibe.ch/svn/pinocchio@896 475e7f6d-5e0e-419c-b4a3-23a65b5ccdbc

+++ p2/system/ast/Variable.h
@@ -6,10 +6,10 @@
 /* ========================================================================= */
 
 struct AST_Variable_t {
-    unsigned int    index;
-    Object          key;
-    Object          name;
-    AST_Info        info;
+    unsigned long int   index;
+    Object              key;
+    Object              name;
+    AST_Info            info;
 };
 
 CREATE_INITIALIZERS(AST_Variable)
